Fix: Ensure connection flags are supported
The flag parsing in Mysql2::Client first checks if the constant is
defined before setting the flag, and this was failing as the original
client had been swapped out, leading to the odd situation where you
could see the value of the constant, but it was saying it wasn't
defined.
Solution is to also delegate const_defined? to the origin client.

Also bumps the mysql2 version.